requirements.txt
asn1crypto==0.24.0
backports.csv==1.0.7
bcrypt==3.1.7
boto3==1.9.180
botocore==1.12.180
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.2
django-bootstrap4==0.0.7
django-heroku==0.3.1
django-import-export==1.1.0
django-s3-folder-storage==0.5
django-storages==1.7.1
django-tenant-schemas==1.9.0
docutils==0.14
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
jdcal==1.4.1
jmespath==0.9.4
odfpy==1.4.0
openpyxl==2.6.2
paramiko==2.6.0
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.8.3
pycparser==2.19
PyNaCl==1.3.0
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
storage==0.0.4.3
tablib==0.13.0
urllib3==1.25.3
whitenoise==4.1.2
xlrd==1.2.0
xlwt==1.3.0
